login and register functions are done, also changed app.py into modules
modules db.py, routes.py and users.py were made
also made base.html template to use, menubar without logged in user specific items.
