+++ db.py
@@ -1,0 +1,8 @@
+from app import app
+from flask_sqlalchemy import SQLAlchemy
+from os import getenv
+
+
+
+app.config["SQLALCHEMY_DATABASE_URI"] = getenv("DATABASE_URL")
+db = SQLAlchemy(app)
