+++ templates/base_wo_menu.html
@@ -9,19 +9,6 @@
 </head>
 
 <body>
-        <nav class="navbar navbar-expand-lg navbar-light bg-light">
-                <a class="navbar-brand" href="/">HaalarimerkkiDEX</a>
-                <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavAltMarkup" aria-controls="navbarNavAltMarkup" aria-expanded="false" aria-label="Toggle navigation">
-                    <span class="navbar-toggler-icon"></span>
-                </button>
-                <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
-                    <div class="navbar-nav">
-                        <!-- <a class="nav-item nav-link active" href="">Home <span class="sr-only">(current)</span></a> -->
-                        <a class="nav-item nav-link active" href="new">Lisää haalarimerkki</a>
-                        <a class="nav-item nav-link active" href="#">Pricing</a>
-                    </div>
-                </div>
-            </nav>
         <div class="content">
                 {% block content %} {% endblock %}
         </div>
