+++ templates/error.html
@@ -1,0 +1,7 @@
+{% extends 'base.html' %}
+
+{% block content %}
+
+<p>Virhe: {{ message }}</p>
+
+{% endblock %}
