+++ templates/index.html
@@ -1,14 +1,13 @@
-{% extends 'base.html' %}
+{% extends 'base_wo_menu.html' %}
 
 {% block content %}
-    <br>
-    <h1>{% block title %} HaalarimerkkiDEX {% endblock %}</h1>
-    <h1 class="h3 mb-3 font-weight-normal">Please sign in</h1>
-    <label for="inputEmail" class="sr-only">Email address</label>
-    <input type="email" id="inputEmail" class="form-control" placeholder="Email address" required autofocus>
-    <label for="inputPassword" class="sr-only">Password</label>
-    <input type="password" id="inputPassword" class="form-control" placeholder="Password" required>
-    <br>
-    <button class="btn btn-lg btn-primary btn-block" type="submit">Sign in</button>
-
+<br>
+<div style="text-align: center;" class="text-center">
+    <h1>Welcome haalarimerkkiDEX</h1>
+    <div class="btn-group">
+        <a href="/login" class="btn btn-primary btn-important">Sign In</a>
+        <span>&nbsp;</span> <!-- Add a space between the buttons -->
+        <a href="/register" class="btn btn-primary">Register</a>
+    </div>
+</div>
 {% endblock %}
