+++ templates/login.html
@@ -1,0 +1,20 @@
+{% extends 'base.html' %}
+
+{% block content %}
+
+<div class="container text-center">
+    <form action="/login" method="post" class="mx-auto">
+        <div class="mb-3">
+            <label for="username" class="form-label">Tunnus:</label>
+            <input type="text" class="form-control" id="username" name="username">
+        </div>
+        <div class="mb-3">
+            <label for="password" class="form-label">Salasana:</label>
+            <input type="password" class="form-control" id="password" name="password">
+        </div>
+        <button type="submit" class="btn btn-primary">Kirjaudu</button>
+    </form>
+    <p>Jos sinulla ei ole tunnusta, voit luoda uuden tunnuksen <a href="/register">tästä</a></p>
+</div>
+
+{% endblock %}
