+++ templates/new.html
@@ -5,9 +5,9 @@
 
     <b>Tervetuloa</b> <i>sovellukseen</i>!
 
-
     <form action="/send" method="POST">
         <br>
+        <p>Lisääthän mahdollisimman tarkan nimen haalarimerkille!</p>
         <input type="text" class="form-control" id="nimi" name="nimi" placeholder="Haalarimerkin nimi:" required>
         <br>
         <input type="radio" id="tapahtuma" name="hankintapaikka" value="tapahtuma" required>
